+++ typings/rx-document.d.ts
@@ -42,8 +42,8 @@ export declare class RxDocumentBase<RxDocumentType, OrmMethods = {}> {
 
     // attachments
     putAttachment(creator: RxAttachmentCreator): Promise<RxAttachment<RxDocumentType, OrmMethods>>;
-    getAttachment(id: string): Promise<RxAttachment<RxDocumentType, OrmMethods>>;
-    allAttachments(): Promise<RxAttachment<RxDocumentType, OrmMethods>[]>;
+    getAttachment(id: string): RxAttachment<RxDocumentType, OrmMethods>;
+    allAttachments(): RxAttachment<RxDocumentType, OrmMethods>[];
     readonly allAttachments$: Observable<RxAttachment<RxDocumentType, OrmMethods>[]>;
 
     toJSON(): RxDocumentTypeWithRev<RxDocumentType>;